+++ Codeforces/B_Grab_the_Candies.cpp
@@ -8,7 +8,7 @@ using namespace std;
 int main() {
     ios::sync_with_stdio(false);
     cin.tie(0);
-    int t;
+    ll t;
     cin>>t;
     while(t--){
         int n;
